Add clone method in ChipStruct, ChipEventStruct
