+++ scripts/py_matter_idl/BUILD.gn
@@ -44,6 +44,8 @@ pw_python_package("matter_idl") {
     "matter_idl/tests/outputs/global_struct_attribute/jni/DemoClusterClient-InvokeSubscribeImpl.cpp",
     "matter_idl/tests/outputs/optional_argument/jni/MyClusterClient-ReadImpl.cpp",
     "matter_idl/tests/outputs/optional_argument/jni/MyClusterClient-InvokeSubscribeImpl.cpp",
+    "matter_idl/tests/outputs/several_clusters/java/ChipEventStructs.java",
+    "matter_idl/tests/outputs/several_clusters/java/ChipStructs.java",
     "matter_idl/tests/outputs/several_clusters/java/ClusterIDMapping.java",
     "matter_idl/tests/outputs/several_clusters/java/ClusterWriteMapping.java",
     "matter_idl/tests/outputs/several_clusters/jni/FirstClient-ReadImpl.cpp",
