+++ scripts/py_matter_idl/matter_idl/tests/available_tests.yaml
@@ -53,6 +53,8 @@ java-jni:
 java-class:
     inputs/several_clusters.matter:
         java/chip/devicecontroller/cluster/structs/SecondClusterFabricDescriptorStruct.kt: outputs/several_clusters/java/SecondClusterFabricDescriptorStruct.kt
+        java/chip/devicecontroller/ChipEventStructs.java: outputs/several_clusters/java/ChipEventStructs.java
+        java/chip/devicecontroller/ChipStructs.java: outputs/several_clusters/java/ChipStructs.java
         java/chip/devicecontroller/ClusterWriteMapping.java: outputs/several_clusters/java/ClusterWriteMapping.java
         java/chip/devicecontroller/ClusterReadMapping.java: outputs/several_clusters/java/ClusterReadMapping.java
         java/chip/devicecontroller/ClusterIDMapping.java: outputs/several_clusters/java/ClusterIDMapping.java
