+++ src/controller/java/BUILD.gn
@@ -422,6 +422,8 @@ android_library("java") {
 
   if (matter_enable_java_generated_api) {
     sources += [
+      "generated/java/chip/devicecontroller/ChipEventStructs.java",
+      "generated/java/chip/devicecontroller/ChipStructs.java",
       "generated/java/chip/devicecontroller/ClusterReadMapping.java",
       "generated/java/chip/devicecontroller/ClusterWriteMapping.java",
       "src/chip/clusterinfo/ClusterCommandCallback.java",
@@ -432,8 +434,6 @@ android_library("java") {
       "src/chip/clusterinfo/InteractionInfo.java",
       "src/chip/devicecontroller/ChipTLVValueDecoder.java",
       "zap-generated/chip/devicecontroller/ChipClusters.java",
-      "zap-generated/chip/devicecontroller/ChipEventStructs.java",
-      "zap-generated/chip/devicecontroller/ChipStructs.java",
       "zap-generated/chip/devicecontroller/ClusterInfoMapping.java",
     ]
   }
